third_party/blink/web_tests/virtual/css-gap-decorations/external/wpt/css/css-multicol/parsing/column-rule-width-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['column-rule-width'] = "medium 40px" should not set the property value
  assert_equals: expected "" but got "medium 40px"
Harness: the test ran to completion.
